// src/infrastructure.timetable-generator/proto/generated/core/timetable_grid_schedule.ts
export interface TimetableGridSchedule {
  /** date-time */
  date: string;
  diaryId: string;
  teacherId: string;
  groupId: string;
  timeSlot?: TimeSlot | undefined;
  roomId?: string | undefined;
}

function createBaseTimetableGridSchedule(): TimetableGridSchedule {
  return {
    date: "",
    diaryId: "",
    teacherId: "",
    groupId: "",
    timeSlot: undefined,
    roomId: undefined,
  };
}

export const TimetableGridSchedule: MessageFns<TimetableGridSchedule> = {
  encode(message: TimetableGridSchedule, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.date !== "") {
      writer.uint32(10).string(message.date);
    }
    if (message.diaryId !== "") {
      writer.uint32(18).string(message.diaryId);
    }
    if (message.teacherId !== "") {
      writer.uint32(26).string(message.teacherId);
    }
    if (message.groupId !== "") {
      writer.uint32(34).string(message.groupId);
    }
    if (message.timeSlot !== undefined) {
      TimeSlot.encode(message.timeSlot, writer.uint32(42).fork()).join();
    }
    if (message.roomId !== undefined) {
      writer.uint32(50).string(message.roomId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): TimetableGridSchedule {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseTimetableGridSchedule();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.date = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.diaryId = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.teacherId = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.groupId = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.timeSlot = TimeSlot.decode(reader, reader.uint32());
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.roomId = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): TimetableGridSchedule {
    return {
      date: isSet(object.date) ? globalThis.String(object.date) : "",
      diaryId: isSet(object.diaryId)
        ? globalThis.String(object.diaryId)
        : isSet(object.diary_id)
          ? globalThis.String(object.diary_id)
          : "",
      teacherId: isSet(object.teacherId)
        ? globalThis.String(object.teacherId)
        : isSet(object.teacher_id)
          ? globalThis.String(object.teacher_id)
          : "",
      groupId: isSet(object.groupId)
        ? globalThis.String(object.groupId)
        : isSet(object.group_id)
          ? globalThis.String(object.group_id)
          : "",
      timeSlot: isSet(object.timeSlot)
        ? TimeSlot.fromJSON(object.timeSlot)
        : isSet(object.time_slot)
          ? TimeSlot.fromJSON(object.time_slot)
          : undefined,
      roomId: isSet(object.roomId)
        ? globalThis.String(object.roomId)
        : isSet(object.room_id)
          ? globalThis.String(object.room_id)
          : undefined,
    };
  },

  toJSON(message: TimetableGridSchedule): unknown {
    const obj: any = {};
    if (message.date !== "") {
      obj.date = message.date;
    }
    if (message.diaryId !== "") {
      obj.diaryId = message.diaryId;
    }
    if (message.teacherId !== "") {
      obj.teacherId = message.teacherId;
    }
    if (message.groupId !== "") {
      obj.groupId = message.groupId;
    }
    if (message.timeSlot !== undefined) {
      obj.timeSlot = TimeSlot.toJSON(message.timeSlot);
    }
    if (message.roomId !== undefined) {
      obj.roomId = message.roomId;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<TimetableGridSchedule>, I>>(base?: I): TimetableGridSchedule {
    return TimetableGridSchedule.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<TimetableGridSchedule>, I>>(
    object: I,
  ): TimetableGridSchedule {
    const message = createBaseTimetableGridSchedule();
    message.date = object.date ?? "";
    message.diaryId = object.diaryId ?? "";
    message.teacherId = object.teacherId ?? "";
    message.groupId = object.groupId ?? "";
    message.timeSlot =
      object.timeSlot !== undefined && object.timeSlot !== null
        ? TimeSlot.fromPartial(object.timeSlot)
        : undefined;
    message.roomId = object.roomId ?? undefined;
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin
  ? T
  : T extends globalThis.Array<infer U>
    ? globalThis.Array<DeepPartial<U>>
    : T extends ReadonlyArray<infer U>
      ? ReadonlyArray<DeepPartial<U>>
      : T extends {}
        ? { [K in keyof T]?: DeepPartial<T[K]> }
        : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin
  ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}