// src/infrastructure.timetable-generator/proto/generated/service/service_generate_response.ts
export interface ServiceGenerateResponse {
  requestId: string;
  isSuccessful: boolean;
  resultSuccess?: ServiceGenerateResponseResultSuccess | undefined;
  resultError?: ServiceGenerateResponseResultError | undefined;
  dateTimeIssued: string;
}